upwork-agent/requirements.txt
openai>=1.41.0
faiss-cpu>=1.8.0
numpy>=1.26
jinja2>=3.1
pydantic>=2.8
python-dotenv>=1.0
fastapi>=0.112
uvicorn>=0.30
streamlit>=1.36